Create EnableEntity.ts
`EnableEntity.ts` is a new file that contains function named processEntity that `RevisionRecordController.ts` passes entity related data to. This function will process the entities, as the code within the function is utilized by every entity (outcomes, tasks, patients, contacts, careplans) and does not need to be duplicated under each case within the `RevisionRecordController.ts` switch statement. The _params_ feature of typescript was used in the processEntity function, and accepts two arguments. The entities that are iterated over by a for loop, and the {entity}Repository which is a constant. If the record or entity exists already, a 409 response is sent, which works the same as the previous setup when just outcomes and tasks were enabled.